Fix array_keys($array, null) to make it equivalent to PHP.
Currently in Hiphop "array_keys($data)" and "array_keys($data, null)"
are equivalent, as the optional second parameter has a default value of
null. In PHP instead they are different, with the second one in
particular returning only the keys from $data where the corresponding
value is null.

The following patch adds a special "not_given_variant" global variable
which can be used to distinguish the case where a parameter is provided
with the value null from the case where no parameter is provided at
all.