Use better escaping for identifiers.
There are currently a few different implementations of string escaping
with the purpose of generating valid C++ identifiers, which are used in
different places. Some of them don't correctly escape all the
characters which are invalid in identifiers, and can also generate the
same identifier out of different original strings, which could result
in collisions. For example these problems affect the compiled templates
generated by Smarty, which use characters like % or ^ in the filenames.

The best implementation for identifier escaping seems to be
CodeGenerator::formatLabel. This patch extracts that function to Util
and removes the other implementations. To improve readability, some
shorter escape codes have also been introduced for common characters
(for example "." -> "$_"). The patch also removes the LabelEscape
option for simplicity and correctness, since guaranteeing that no
collisions are possible becomes more complicated if the escaping string
is not fixed in advance.