Don't remove blank text elements in the soap extension
The soap extension assumes that text elements containing only
whitespaces in the response are insignificant and removes them before
converting the xml into a php representation. This assumption is not
correct in general, for example in case a cryptographic signature must
be generated based on the content of the text elements.

In general situations where blank text elements are indeed
insignificant can be specified in the wsdl by using xsd:token instead
of xsd:string for the type of the text element. These cases are already
correctly handled in the extension and the whitespace is trimmed later
on in the flow.

This patch adds a parameter in the functions used to load an xml
document inside the soap extension to specify whether to remove blank
text elements or not. Removal of these elements is then disabled for
general soap requests/response parsing, but left enabled for
wsdl/schema parsing because of other code relying on it.